requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
decorator==4.4.2
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
moviepy==1.0.3
numpy==1.25.1
Pillow==9.5.0
pillow-heif==0.12.0
proglog==0.1.10
requests==2.31.0
tqdm==4.65.0
urllib3==2.0.4
